requirements.txt
requests
python-dotenv
qrcode
Pillow
pynput